requirements.txt
# Birthday Bot - Система уведомлений о днях рождения
# С уникальной генерацией поздравлений

python-telegram-bot>=20.0    # Для работы с Telegram API
pandas>=2.0.0                # Для обработки Excel файлов
openpyxl>=3.0.0              # Для чтения .xlsx файлов
APScheduler>=3.10.0          # Для планирования уведомлений
python-dotenv>=1.0.0         # Для управления переменными окружения

# Особенности системы:
# - Автоматическая генерация уникальных поздравлений
# - Компонентный подход (3-5 предложений на поздравление)
# - Учет пола и сезона
# - Автоматический подбор эмодзи
# - Коллективные поздравления для нескольких именинников
